pyproject.toml
[project]
name = "vad-genai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.19.0",
    "google-generativeai>=0.8.5",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "ultralytics>=8.3.153",
]
